Move Trace macros to new header to avoid naming conflicts
